Check termination on erased terms, unrestricted term primitive new syntax for operators
Previously, Formality disabled stratification checks on erased terms.
That was problematic, because, then, they couldn't be interpreted as
expected. For example, one could expect that the function below could
only return 0:

    foo : {x : Word, ~e : x == 0} -> Word
      x

But this isn't true, since we could use an unhalting term as `e`:

    bad : {P : Type} -> P
      bad(P)

    main
      foo(7, ~bad(7 == 0))

This commit fixes this issue by enabling stratification checks on erased
terms, making them terminating. The problem with this approach is that
it makes the language considerably less expressive. For example, we're
not able to implement inductive Nats anymore, since it uses its argument
twice, one time on the erased position.

To solve this issue, this commit introduces a new primitive for
unrestricted terms:

    syntax | meaning
    ------ | ---------------------------------------
    -A     | `A` is an unrestricted term of type `A`
    %t     | Converts `t` to an unrestricted term
    +t     | Converts `t` back to a restricted term

Semantics-wise, an unrestricted term of type `A` is either a well-typed
term of type `A`, or a never-ending computation. In other words, this
separates the language in two fragments, a non-terminating one, which
can be used for types and programs, and a terminating one, which can be
used for proofs.

This makes the language considerably more powerful. It allows us not
only to implement inductive Nats again:

    Ind : Type
      ${self}
      {~P     : {r : Ind} -> Type} ->
      {step   : ! {~r : -Ind, i : P(+r)} -> P(step(+r))} ->
      ! {base : P(base)} ->
        P(self)

    step : {r : Ind} -> Ind
      new(~Ind){~P, step}
      dup step = step
      dup func = (use(r))(~{x} $ P(x), #step)
      # {base}
        step(~%r, func(base))

    base : Ind
      new(~Ind){~P, step}
      dup step = step
      # {base}
        base

But it can also be used to create Haskell-like infinite structures:

    T List {A : Type}
    | cons {h : A, t : List(A)}
    | nil

    ones : -List(Word)
      % cons(~Word, 1, +ones)

And even implement familiar recursive algorithms:

    T Nat
    | succ {pred : Nat}
    | zero

    double : {case x : Nat} -> -Nat
    | succ => % succ(succ(+double(x.pred)))
    | zero => % zero

Of course, unrestricted function must return unrestricted terms:

    main : -Nat
      double(0n3)

The cool thing about this setup is that, while we're able to represent
never-ending computations and, thus, "prove" paradoxes:

    main : -Empty
      %+main

Those paradoxes are annotated with a `-`, which essentially tells us
they can't be interpreted as mathematical proofs. On the other hands, if
the type of a term is not annotated with a `-`, then this term can be
interpreted as a mathematical proof, even if the type itself mentions
unrestricted programs. For example, this is a perfectly valid proof that
the double of 2 is 4:

    main : 0n4 == +double(0n2)
      refl(~0n4)

Because, while `double` isn't terminating, the proof itself is. In other
words, we can now implement and prove theorems about recursive programs,
as long as those proofs are terminating. This very cool idea was
proposed on the Zombie paper, and it marries very well with Formality.

This commit also changes the syntax of numeric operators as follows:

    op   | name
    ---- | ----
    .+.  | addition
    .-.  | subtraction
    .*.  | multiplication
    ./.  | division
    .%.  | modulus
    .^.  | exponentiation
    .&.  | bitwise and
    .|.  | bitwise or
    .#.  | bitwise xor
    .!.  | bitwise not
    .>>. | bitwise shift right
    .<<. | bitwise shift left
    .>.  | greater-than
    .<.  | less-than
    .==. | equals
    .++. | float addition
    .--. | float subtraction
    .**. | float multiplication
    .//. | float division
    .%%. | float modulus
    .^^. | float exponentiation
    .f.  | uint to float
    .u.  | float to uint

This commit also changes the syntax of type-level identity to:

    A :: Type

This commit also changes the syntax of unboxing to:

    $t