Type-checker returns algorigth. Auto-fill tilde on applications
Since the type-checker is now capable of filling holes, that caused an
issue: what if you attempted to reduce a term with a hole? For example,
the program below:

    main : {P : Bool -> Type, f : {x : Bool} -> P(x)} -> P(true)
      f(?a)

Would fill the hole `?a` as `true`, causing the program to pass the
type-checker. But if you attempted to reduce it, you'd get:

    {P, f} => f(?a)

And compiling to inets would be impossible, since there is a hole there.
To solve this problem, the type-checker now returns an erased version of
your program with the holes filled.

That change also allows us to improve the language by auto-filling
erasure annotations. For example, the program below:

    main : {f : {~x : Bool} -> Bool} -> Bool
      f(true)

Would be a type-error ("mismatched erasure"), because the user forgot to
annotate a `~` on the `f(true)` application (which is important to
inform the compiler to discard the `true` argument). Since now the
type-checker returns the algorithm, it can just auto-fill those erasure
annotations for you, so the program above now type-checks and the user
never needs to use `~` on applications anymore.