Temporarily disable all auto-fillings
The previous commit introduced a simplistic form of auto-filling for a
bunch of symbols and holes. While this was cool and handy, the way it
was implemented wasn't very robust and 1. caused some inconveniences
(such as badly-filled holes causing type errors, preventing them being
used for what they were originally introduced), 2. wasn't very
extensible, in the sense it can't improve much further. 3. it slowed
down the type-checker slightly, but if we wanted typed hole filling, the
slowdown would be considerable.

We could have way more powerful auto-fillings that performed proper
unification on holes and that added all missing boxing, type-identity
and restriction annotations, but that would make the core bigger and
slower. Those things should probably be external to core, which is meant
to be a simple / dumb proof checker and not evolve into a complex proof
assistant. As such, I'm removing those from core for now, as I think on
the best way to separate those concerns.