Separate halt and type checker, support recursion on inets
Now Formality will allow potentially non-terminating terms to pass the
type-checker. A separate test, haltcheck, can be used to determine if
the term halts. Non-halting terms may be inhabited by bottom and, thus,
can't be seen as mathematical proofs. Since well-typed terms are
elementary affine, the only way they can fail to halt is by direct
recursion. The haltchecker is, right now, too conservative: it will
exclude any kind of recursion. In a future, it might be improved to
allow well-founded recursion. When a term succeeds to type-check but
fails to halt-check, the CLI will show a warning sign, symbolizing that
it is correct *if* it fails to halt.

This allows the user to implement Agda-like, well-founded recursive
algorithms. This is much simpler than the Ind-based recursion approach.
As an example, suppose we had a type of Nats with a statically known
length (i.e., like Vectors without elements):

    T Cat (n : Nat)
    | csucc {~s : -Nat, pred : Cat(s)} (succ(s))
    | czero (zero)

This is what we had to write to double it using Ind-based recursion:

    #mul2*N : !{n : Cat($nat(N))} -> Cat($nat(twice(N)))
      case/Cat n
      + refl(~Nat, ~succ($nat(N))) as e : Equal(Nat, n.n, succ($nat(N)))
      | csucc =>
        let e = cong(~Nat, ~Nat, ~succ(n.s), ~succ($nat(N)), ~pred, e)
        let p = rewrite(~Nat, ~n.s, ~$nat(N), e, ~Cat, n.pred)
        csucc(~succ($nat(twice(N))), csucc(~$nat(twice(N)), mul2(p)))
      | czero => absurd(zero_isnt_succ(~$nat(N), e), ~Cat($nat(twice(step(N)))))
      : Cat($nat(twice(step(N))))
    halt: czero

Not only this has an weaker API since it involves boxes, but it also
requires a lot of auxiliary proofs, because it isn't a direct recursion
but rather an emulation of it through Nat induction. With the Agda-like
well-founded recursion, the same function can be written as:

    mul2b : {~s : -Nat, n : Cat(s)} -> Cat(dbl(s))
      case/Cat n
      | csucc => csucc(~succ(dbl(n.s)), csucc(~dbl(n.s), mul2b(~n.s, n.pred)))
      | czero => czero
      : Cat(dbl(n.n))

Which is considerably simpler. Also, note that, previously, the
recursion syntax couldn't handle branching recursion, which works now.
In a future, perhaps the Ind-recursion syntax should be removed and
boxes should be considered advanced features, so users will box and
unbox manually. So, depending on how things go, Ind-recursion syntax
and boxed definitions should be considered deprecated soon.

Note that `mul2b` is applied to a structurally smaller value, so it
terminates. The type checker is *not* capable of identifying that this
function is safe yet, so it presents a warning, which must be
interpreted as "this term is a valid *if* it terminates". So, for
example, one can create paradoxes such as:

    main : Empty
      main

With will, too, be accepted with a warning. In a future, the
type-checker should be able to separate those paradoxes from
well-founded recursive algorithms using similar algorithms as Agda, but
adapted for λ-encodings. For now, proofs with a "warning" status must be
taken with a grain of salt and not accepted in a decentralized setting.

Fortunatelly (and perhaps surprisingly to some), optimal reduction is
compatible with arbitrary recursive algorithms! What really causes a
problem on interaction nets is when you have "nested duplications";
things like `({f} f(f(x)))({f} f(f(x)))`. When compiling recursive
functions to interaction nets, fan nodes point downwards, towards a
closed term, rather than upwards, towards a λ-bound variable. That makes
it free of those nested duplications and, thus, sound. This commit
implements sound compilation of arbitrary recursive functions to inets.