Major syntax changes towards well-founded recursion
:: Boxed and ind-recursive definitions removed

Since we're pursuing the well-founded recursion route, boxed and
ind-based recursive definitions were removed from fm-lang. That removes
a lot of "auto-magic" from it, making it considerably simpler. This also
removes related features like ind application (func*N), auto-unboxing
and so on.

:: Case arguments removed

They were implemented in a not very robust manner and were incomplete
(didn't support indexed datatypes). After the move feature, they are
less necessary. This simplifies fm-lang further.

:: Case expressions don't need the datatype

Now you're able to write just `case b` instead of `case/Bool b`.

:: Strings now are just lists of codepoint chars

After the change of Num, this seems like the most logical step.

:: Syntax for numeric literals changed

Now, 1234n creates a Nat, and 1234i creates an Ind. This looks cleaner
than the old syntaxes, 0n1234 and *1234 respectively.

:: Syntax for functions and foralls changed

Now, lambdas and foralls use the `(x : A) => x` and (x : A) -> B`
syntaxes, respectively.

:: Top-level definition syntax changed

Now, instead of:

    main : {x : A, y : B} -> C
      ...body...

Top-level definitions are written as:

    main(x : A, y : B) : C
      ...body...

:: Datatype parameter syntax changed

Now, instead of:

    T Vector {A : Type} (len : -Nat)
    | vnil                                                 (zero)
    | vcons {~len : -Nat, head : A, tail : Vector(A, len)} (succ(len))

A datatype is written as:

    T Vector<A : Type> (len : -Nat)
    | vnil                                                : Vector(A, zero)
    | vcons(~len : -Nat, head : A, tail : Vector(A, len)) : Vector(A, succ(len))

:: The type of Type typed variables can be omitted from foralls

For example, `(~A : Type) -> ...` can be written as `(~A) -> ...`.