Formality is now JavaScript
This fixes a small bug where many spaces would end up on the tokens
object returned by the parser. This also makes the do-statement more
JS-like. For example, this is the new parse_nat function:

    parse_nat : Parser(Nat) {
      var chr = next;
      if chr .==. 'S' then {
        var pred = parse_nat;
        return succ(pred);
      } else {
        return zero;
      }
    }

Of course, this is translated to a series of binds:

    parse_nat : Parser(Nat)
      bind(__ next, (chr) =>
      if chr .==. 'S' then
        bind(__ parse_nat, (pred) =>
        return(_ succ(pred)))
      else
        return(_ zero))

In a future, we might add for-loops and other control constructs.