Major cleanup towards feature freeze and 1.0 (WIP)
This commit cleans up a lot of features and experimental code. This
includes the removal of pairs and numbers as features of the language,
because we can't decide the best and most future-proof way to implement
them, and also to make the language even simpler. This will be revisited
in a future. This makes Formality target language exactly, precisely the
affine λ-calculus with well-founded recursion.

There is still a lot to clean up and improve, so this is a WIP. The
documentation is also up for a revamp and is probably outdated in many
places.