Separate Type, Stratification and Termination checkers
Short explanation:

The more I think about it, the more I believe termination checker and
box inferencer should be separate from the base language, so I'll start
arranging things in that direction.

Long explanation:

The main reason the new runtime it is so ridiculously faster than the
previous interaction net evaluator is that 1. it removes redundant
pointers by giving a order to the graph, 2. it performs copies of
references in chunks rather than incrementally. This gives us a ~40x
speedup and, IMO, shows how much long-term potential this language has.

Sadly, the new runtime doesn't feature fan nodes. My initial reaction
was to just get rid of boxes for good. They are complex, unliked, and we
barely used them anyway. But this means we also throw away lazy
duplications, optimal sharing characteristics, O(log(N)) function
exponentiation, runtime fusion and all those "cool things". Without
them, Formality will be just a very fast affine language. Which isn't a
bad, we could live with that, it is just a different design. But,
thinking about it, I want to save those things.

I've spent the last 2 days trying to figure out if I can extend the new
runtime with duplications. My conclusion is that it absolutely doable,
but keeping pointers correct is complex. It requires further work to
either 1. add a new buffer to keep track of matching variables, or 2.
figure out a clever evaluation order on which pointers don't misalign.
Option 1 is easy to do, but will probably be slower. I'm not sure option
2 is possible. A potential approach is to try it the other way around:
instead of making a λ-evaluator inspired on inets, optimize inets to not
need duplicate pointers. That will make the correct evaluation order
easier to implement and debug.

Even if we manage to add duplications on the new runtime, boxes are
still an unintuitive feature. So, for the sake of user-friendliness, I
think I'll just remove the linearity checks of the language, entirely.
That is a huge change, as it means it will not be safe to run well-typed
Formality terms on interaction nets. But perhaps this is actually fine.
After all, we can add box inference to decide if a program is compatible
with optimal reductions in a future (Daniel's work). If it is not, we
can just evaluate it more traditionally.

So, that changes the way we view Formality. Previously, we've made a
huge effort to implement "optimal proof language". The problem is that
those two goals are conflicting. Proofs require consistency, which
requires a set of limitations that are not desirable for someone that
just wants to write programs. Similarly, beta-optimality requires a set
of limitations that are unecessary for someone that just wants proofs,
or doesn't care about performance.

As such, Formality base-layer should be just a dependently typed
programming language. Then, with supplementary termination and
stratification checks, it can be used as either a proof language or as a
beta-optimal language, depending on the user's needs, but not
mandatorily so. This makes it more flexible and, mostly, allows us to
pospone those matters and focus in other aspects of the language, such
as applications and ecossystem. We spent too much time trying to find a
perfect subset that seems to not exist.

Basically:

- Type-Checker: ensures there won't be runtime errors, segfaults.

- Termination-Checker: ensures the program will terminate and, thus, can
be seen as a valid mathematical proof.

- Stratification-Checker: ensures the program can be reduced with the
abstract fragment of Lamping's optimal reduction algorithm, that is,
without book-keeping machinery (also known as the Oracle). This allows
it to be reduced optimally (i.e., with the least amount of
beta-reductions and copying), AND efficiently (i.e., without a huge,
constant overhead that makes it unpractical).

The main benefit of this is making the language more flexible. For
example, the old language didn't allow structural recursion, which made
it extremelly complex to work with recursive datatypes like Nats and
Lists, as we needed additional Church-encoded datatypes to fold over
them. Theorem proving was even harder, since we needed to use the Church
structure to bound the size of the recursed value, in order to prove
that the base case was unreachable, essentially re-implementing
well-founded recursion on every proof involving Vectors, Fins or similar.

Similarly, while the old language could be soundly evaluated on the
interactio net runtime, there were many safe programs it excluded. For
example, copying a Nat can be done as:

    copy(n : Nat) : Pair(Nat, Nat)
      pair(__ n, n)

But the old language didn't allow that program. Arguably, a separate
checker will be more flexible than the old box system, allowing a bigger
class of programs, and not requiring additional effort from the user.
This, to me, seems like the best choice on the long term, so I'm
adjusting the language towards this direction.