Initial specification
Since we're moving away from having boxes, affinity, elementarity and
termination as core language features, in favor of separate
complexity-class checkers, Formality is really simple and stable right
now.

I think we should freeze it roughly as it is and document both its core
AST and its high-level syntax in a single, small paper with the goal of
becoming a standard reference for implementors. I'd also like to stop
separating the core AST from the syntax, the sugars aren't that complex
and I think we can include them as part of the whitepaper.

Regarding compilation, we can briefly comment how the linear and
elementary affine subset can be compiled to efficient runtimes such as
interaction combinators, but a full description of those runtimes should
be in a separate, i.e., those aren't to be seen as parts of the
language. Instead, Formality should be seen as a simple proof language
which makes it very easy to work on those "fast" subsets if you want to,
but that doesn't restrict you if you want to use it like a Haskell.

So, in short, this paper should just document the base language as it
is, including lambdas, self-types, numbers etc. The language is still
missing uints and some minor things, but it is mostly stable as is, so
it is a good moment to start the paper.