Amazing type-guided JS compiler
The command `fmcjs file.fmc` now compiles a Formality-Core to JavaScript
using type information. Common datatypes are compiled to native values:

- Bool is compiled to native Bool

- Nat is compiled to native Bignum

- U32 is compiled to native Number

- String is compiled to native String

To trigger that, you must comment `//prim//` after the datatype's type:

    Nat: Type //prim//
      self<P: Nat -> Type> ->
      (zero: P(zero)) ->
      (succ: (pred: Nat) -> P(succ(pred))) ->
      P(self)

This works by converting datatypes to native values when they are
constructed with a lambda, and performing the right elimination when
they are eliminated with an application. Without type information, that
would be impossible. To understand, consider the Nat datatype.

1. If we simply replaced the `zero`/`succ` constructors to create native JS
  numbers, the user could still build a λ-encoded Nat without using
  them, so some Nats would be numbers and others would be lambdas,
  resulting in bugs.

2. When pattern-matching on a Nat represented as a number, it would result in
  something like `72(case_zero)(case_true)`, which is a JS error since
  you can'y apply a number like a function.

This solution inserts instantiators and eliminators in the right
locations, allowing lambdas to be converted native encodings when they
are constructed, and pattern-matching correctly when they are eliminated.

With types being replaced by native numbers, we are also able to replace
certain functions by native ones. For example, `add` is replaced by `+`.
To trigger it, you must comment `//prim//` after the function's type:

    nat_add: Nat -> Nat -> Nat //prim//
      (n) (m)
      n<() Nat>
      | m;
      | (n.pred) nat_add(n.pred)(succ(m));

This is very good, since it increases the performance by several orders
of magnitude without needing those types to be part of Formality-Core.

The compiler can also perform tail-call optimization, which is
extremelly important to allow stack blowups in JavaScript. It can be
enabled commenting `//loop//` after a function's type:

    nat_sum_go: Nat -> Nat -> Nat //loop//
      (n) (r)
      n<() Nat>
      | r;
      | (n.pred) nat_sum_go(n.pred)(nat_add(n)(r));

Note this isn't checked, so using //loop// in non-tail-call optimizable
functions will result in errors.