String literals sugar
Formality-Core is intended to be used as an interchange format for
algorithms and proofs. It has two problems, though:

1. Its syntax is too simple and limited to be practical. The plan was to
  transpile a more friendly syntax to it, but, if we distributed
  Formaltiy-Core files, we'd lose that syntax, making that useless.

2. There is no efficient way to embed raw data in Formality-Core files.
  That means strings, numbers, images would get huge in code.

To address those issues, we're adding a syntax sugar for string
literals. Thanks to dependent types, we can use those to have "first
class syntax sugars". For example:

    parse_nat : (str : String) -> ... t(str) ...
      (str) ... f(str)

    main : Nat
      Nat.double(parse_nat("123"))

Here, `parse_nat` returns a `Nat` if "123" is a valid string, or `Unit`
if it isn't, resulting in a type error. Any syntax sugar of a
hypothetical higher level language could be internalized through string
literals, so this might be a better solution.

Moreover, string literals allow us to embed data efficiently on
Formality-Core source files by, for example, using a Base64 encoding.

While this goes against the spirit of Formality-Core, it is
simple enough not to increase its complexity significantly, and the
benefits are very significant.