High-order type checking and simpler equality
The type-checker now operates on high-order terms directly. Previously,
the evaluator used HOAS, but the type-checker used low-order. This
required a full conversion on every call. Now type-checking is 100% HOAS
and shares the same terms with the normalizer. This is only possible due
to the `Ann` constructor, which acts as a bridge between both. The
implementation doesn't even need low order terms at all, shift/subst
were completely removed. I never heard of this being done before and the
results are great.

Equality now doesn't use UnionFind. Instead, it just returns true when
the hash of the two terms are equal, or when we re-visit an equation.

Terms uses Bruijn levels instead of Bruijn indices. This removes the
need for `bind_free_vars`, and simplifies other functions.

This speeds up type-checking of moonad/moonad by 5x, and should decrease
the actual complexity of the language considerably. A further improvement
would be to remove low-order terms entirely, making the parser return
high-order terms.

Bruijn level bugfixes