Update compiler to new HOAS format
The compiler was still using the old format, which required having an
old version of the language and converting. It was also very messy,
having copies of the same logic on JS/HS. This updates the compiler to
the new HOAS format and cleans up the code considerably by first
compiling Core to an intermediate language called Comp, and then
compiling to the final target. Haskell was temporarily disabled (TODO).