Fix subtle bug on typesynth
Typesynth defines `defs[name].core = null` when it starts type-checking
a term. This allows recursive occurrences of non-synthetized terms to
halt, giving it a chance to synthetize holes. The problem is that
typesynth would side-effectively leave `defs[name].core = null` even
when it failed to type-check. This could cause the following situation:

- Term A and C depends on term B

- typesynth(A) attempts to typesynth B, but it fails due to an undefined ref, X

- Formality loader loads the missing ref, X

- typesynth(A) now succeeds. But, since defs[B].core is null, it doesn't
  call typesynth(B) inside typeinfer. This causes defs[B].core to stay
  null. If defs[B].core was undefined, though, that call of typesynth(A)
  would call typesynth(B), which would succeed.

- typesynth(C) fails because defs[B].core is null, and thus its holes
  aren't filled, even though all the required definitions for
  defs[B].core to succeed are available now. If typesynth(B) was called
  before typesynth(A), typesynth(C) would succeed instead.

In other words, this causes a situation where the same term `C` may fail
or succeed to type check depending on the order you call typesynth on
other terms. In order to fix it, typesynth must clear the `defs[X].core = null`
definition when it fails.