Namespacing fixes, remove complex hole search
This fixes some namespacing issues with the previous commit.

This also removes complex search on the hole filling algorithm. Now, it
will only fill immediate holes. This means the algorithm won't be able
to do neat things like solving quadratic equations at the type level in
a prolog-like fashion. But it makes the implementation much simpler,
faster, and covers all the important cases. In fact it covers more
cases, since, previously, we weren't able to, for example, fill the
holes of the Sigma.new constructor, since the compiler would get stuck, even
though Sigma.new is fillable with just immediates.