Fix patch on def bug
Patching on defs was buggy because the type checker expanded defs,
causing path info to be lost. In order to fix it, the type checker now
checks defs similarly to let. The problem is that this would cause defs
used to create lambdas to cant_infer error. To solve that, now the
cant_infer error will instead add a hole to the term. This also now
allows lets to be used to define local lambdas (but not type synonyms).