add function to delete node in GMap
Adds a function that receives a key, a GMap, and a comparator function and deletes the node with that key. The node is found recursively; when found, his content will be replaced by: nothing if has no children; left tree, if right is empty; minimun node on the right side, if right is not empty. When the recursion goes back to the root of the tree, all of the nodes is rebalanced if needed.