Do-Notation desugars to Block type
The Block type is defined as:

use Base/Block/ as BK/

BK/Block
: ∀(M: ∀(A: *) *)
  ∀(R: *)
  *
= λM λR #[]{
  #Ask{ A:* query:(M A) then:(∀(a: A) (BK/Block M R)) }: (BK/Block M R)
  #Ret{ result: R }: (BK/Block M R)
}

It works like a List, except:
- Elements are wrapped in a type constructor
- The contained elements are heterogeneous
- There is a λ-binder between each 'cons' and its tail

This element essentially captures the concept of a do-block, except as a
first-class construct of the Kind language, allowing us to represent
monadic computations as pure data, just like free monads, but without
the functor overhead.

Now, the do-notation will desugar to a block (i.e., a series of Ask/Ret
constructors), instead of dsguring directly to a series of bind/pure
calls. We can then run a block by using Block/run:

use Base/Block/ as BK/
use Base/Maybe/ as MB/
use Base/Trait/ as T/

BK/run
: ∀(M: ∀(x: *) *)
  ∀(m: (T/Monad M))
  ∀(T: *)
  ∀(b: (BK/Block M T))
  (M T)
= λM λ{
  #Monad: λbind λpure λT λ{
    #Ask: λA λquery λthen (bind A T query λanswer (BK/run M #Monad{bind pure} T (then answer)))
    #Ret: λresult (pure T result)
  }
}

The main advantage of this approach is that now, the return type of the
block is fixed, which allows the bidirectional type checker to infer the
return type of a do-notation block:

use Base/Block/ as BK/
use Base/Maybe/ as MB/
use Base/Trait/ as T/

Main
: (MB/Maybe U32)
= (MB/run _ do {
    ask x = #Some{50}
    ask y = #Some{20}
    ret ?here
  })

Previously, Kind wouldn't be able to infer that `?here : U32`. This will
be useful for when we want to interrupt a do-block with a case-of
(although not implemented yet), since we need that type information.