Ignore redundant cases of match in type checking
Now, when writing pattern matching lambdas, if a constructor is passed
more than once or a default case is given to a complete match, it is
simply ignored during type checking.

For example, this now type checks:

Bool/not : Bool -> Bool
| #True  = #False
| #False = #True
| #False = ?unreachable
| _      = ?unreachable

This has the downside of not rejecting incorrect terms if they are
redundant.

This is being introduced so that the default cases of the flattening is
ignored when the function is already complete. Previously this caused
checking errors because when the default rule is redundant it may be
missing some cases.