Use clearer var names when flattening, fix wrong name being generated
Now the generated lambdas are always (hopefully) correctly named, with
matching variable uses and proper reconstruction of the matched
constructor.

We also use 'use' to reconstruct the constructors instead of 'let', since if the
generated variable is not used it can sometimes not have enough
information for typing, leading to a type error.

The following function:

Bits/test : Bits -> Bits -> Bits
| #O{#O{a2}} #E    = a2
| a0         #E    = a0
| #O{a1}     #O{_} = a1
| a0         b     = b

Now desugars to:

λ{
  #O: λ{
    #O: λa2 λ{
      #E: a2
      #O: λb.0 use a1 = #O{a2} a1
      #_: λb use a1 = #O{a2} use a0 = #O{a1} b
    }
    #_: λa1 λ{
      #E: use a0 = #O{a1} a0
      #O: λb.0 a1
      #_: λb use a0 = #O{a1} b
    }
  }
  #_: λa0 λ{
    #E: a0
    #_: λb b
  }
}

We could simplify the output further by directly substituting instead of
inserting a 'use'.