Annotate the term returned by check even with suspended checks
When doing a postponed check, instead of replacing the value with `(Met
0 _)`, we generate a fresh metavariable uid, larger than any existing
one to avoid conflicts, and replace the term by `(Met fresh _)`.

After running the postponed check, we insert the annotated term into the
Fill map, using that fresh uid.

After finishing the check, all the remaining metas in the term are
recursively replaced by their filled values.
This also simplifies checking for unfilled metas since we just need to
see if there are any in the substituted annotated term.

This change is being done so that we can use suspended checks when
compiling to JS. Without it it's very annoying to compile since the code
will need a large amount of additional user-provided annotations.